Upload #000002
09-02-2024 :: Release Notes:
* Removed all special portuguese characters from texts and variables.
* Created some new abstractions for current existing functionality
* Implemented some new features on classes of controller logic

+++ src/controller/view/FXMLEquipamentoController.java
@@ -1,6 +1,9 @@
 package controller.view;
 
 import controller.dao.crud.EquipamentoDAO;
+import controller.dao.crud.FuncionarioDAO;
+import controller.dao.crud.utils.EquipamentoUtils;
+import controller.dao.crud.utils.FuncionarioUtils;
 import java.net.URL;
 import java.util.ResourceBundle;
 import javafx.collections.FXCollections;
@@ -20,6 +23,7 @@ import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
 import javax.swing.JOptionPane;
 import model.classe.Equipamento;
+import model.classe.Funcionario;
 
 public class FXMLEquipamentoController {
 
@@ -111,10 +115,11 @@ public class FXMLEquipamentoController {
     private Label lbManutencao;
 
     @FXML
-    private ComboBox<Integer> cbFuncionario;
+    private ComboBox<String> cbFuncionario;
 
     @FXML
     private ComboBox<String> cbManutencao;
+    private ObservableList<String> optionsManutencao = FXCollections.observableArrayList("SIM", "NAO"); // popula o combo box
 
     @FXML
     private Label lbID;
@@ -124,31 +129,39 @@ public class FXMLEquipamentoController {
     
     @FXML
     void handleBtnApagar(ActionEvent event) {
-        if(txtID.getText().isEmpty()) // VERIFICA SE HÁ CLIENTE SELECIONADO
+        if(txtID.getText().isEmpty()) // VERIFICA SE Ha CLIENTE SELECIONADO
         {
-            JOptionPane.showMessageDialog(null, "Não há Equipamento Selecionado!", "Erro!", 0);
+            JOptionPane.showMessageDialog(null, "Nao ha Equipamento Selecionado!", "Erro!", 0);
         }
         
-        Equipamento equipamento = new Equipamento(Integer.parseInt(txtID.getText()));
-        
-        EquipamentoDAO eqDAO = new EquipamentoDAO();
-        eqDAO.delete(equipamento);
-        limpaCampos();
-        readTable();
-        
-        JOptionPane.showMessageDialog(null, "O Equipamento foi apagado com sucesso!");
+        else{
+            int input = JOptionPane.showConfirmDialog(null, "Você tem certeza que quer Apagar este Equipamento?", "Apagar Equipamento", 0, 2);
+            if(input == 0){
+                Equipamento equipamento = new Equipamento(Integer.parseInt(txtID.getText()));
+
+                EquipamentoDAO eqDAO = new EquipamentoDAO();
+                eqDAO.delete(equipamento);
+
+                JOptionPane.showMessageDialog(null, "O Equipamento foi apagado com sucesso!");
+                limpaCampos();
+            }
+            readTable();
+        }
     }
 
     @FXML
     void handleBtnNovo(ActionEvent event) {
         limpaCampos();
+        cbFuncionario.setDisable(true); // DESABILITA O CB EM CASO DE NOVO EQUIP
     }
 
+    // PRECISA DEBUGAR ESSE MeTODO PRA VER PORQUE PAROU DE SALVAR
+    
     @FXML
     void handleBtnSalvar(ActionEvent event) {
-        // VERIFICA SE HÁ CAMPOS VAZIOS CAMPOS PREENCHIDOS
+        // VERIFICA SE Ha CAMPOS VAZIOS CAMPOS PREENCHIDOS
         if(txtNome.getText().isEmpty() || txtMarca.getText().isEmpty() || txtPreco.getText().isEmpty()
-                || txtDataCompra.getText().isEmpty() || txtDataUltimaManutencao.getText().isEmpty() || cbManutencao.getValue().isEmpty()){
+                || txtDataCompra.getText().isEmpty()){
             JOptionPane.showMessageDialog(null, "Todos os campos precisam estar preenchidos", "Erro!", 0);
             return;
         }
@@ -167,10 +180,17 @@ public class FXMLEquipamentoController {
         // ATUALIZA O EQUIPAMENTO SELECIONADO
         else{
             Equipamento equipamento = new Equipamento();
-            EquipamentoDAO eqDAO = new EquipamentoDAO();
-
             equipamento = getDataTableObject();
-            eqDAO.update(equipamento);
+
+            // SE TEM UMA DATA DE MANUTENcaO E O STATUS MANUTENcaO NaO ESTa MUDANDO, ENTaO VAMOS MANTER ELA
+            if(!isManutencaoChangeToYes(equipamento)){
+                EquipamentoDAO eqDAO = new EquipamentoDAO();
+                eqDAO.update(equipamento);
+            }
+            else{ // CASO CONTRaRIO PEGAREMOS A DATA DE HOJE E SETAREMOS NO CAMPO
+                EquipamentoUtils eqDAO = new EquipamentoUtils();
+                eqDAO.update(equipamento);
+            }
 
             JOptionPane.showMessageDialog(null, "Equipamento Alterado com Sucesso!");
         }
@@ -181,10 +201,17 @@ public class FXMLEquipamentoController {
 
     @FXML
     void tableEquipamentosMouseClicked(MouseEvent event) {
+        cbFuncionario.setDisable(false); // HABILITANDO O CB PARA EDIcaO
+        cbManutencao.setDisable(false);
+        
         ObservableList<Equipamento> selectEquipamentos = tableEquipamento.getSelectionModel().getSelectedItems();
         
-        for(Equipamento equipamento : selectEquipamentos) // RECUPERA OS DADOS DA TABELA NOS FIELDS
+        for(Equipamento equipamento : selectEquipamentos){ // RECUPERA OS DADOS DA TABELA NOS FIELDS
             setDataTableObject(equipamento);
+            
+            if (null == equipamento.getDataUltimaManutencao())
+                txtDataUltimaManutencao.setText("<Sem Manutencao>");
+        }
     }
 
     @FXML
@@ -223,14 +250,22 @@ public class FXMLEquipamentoController {
 
      // CARREGA A TABELA
      readTable();
+     readComboBox();
  }
- // initialize é como um FORM_POST_OPEN ou LOAD
- // os métodos de ações estarão aqui pra baixo
+ // initialize e como um FORM_POST_OPEN ou LOAD
+ // os metodos de acoes estarao aqui pra baixo
  
  // CONSTRUTOR
  public FXMLEquipamentoController(){
-     // NÃO PRECISA DO INIATIALIZE, JÁ É AUTOMATICO
+     // NaO PRECISA DO INIATIALIZE, Ja e AUTOMATICO
  }
+ 
+  /***
+  * 
+  * 
+  * AQUI e O SEGUINTE, FALTA FAZER O NOME SELECIONADO NO COMBO BOX ASSIGNAR COM O EQUIPAMENTO ESCOLHIDO
+  * PARA STATUS EM USO DO EQUIPAMENTO
+  */
  
  // INSERE OS DADOS DOS CAMPOS NO DAO E DEPOIS NA TABELA
  public Equipamento getDataTableObject(){
@@ -240,14 +275,31 @@ public class FXMLEquipamentoController {
      equipamento.setMarca(txtMarca.getText());
      equipamento.setPreco(Float.parseFloat(txtPreco.getText()));
      equipamento.setDataCompra(txtDataCompra.getText());
-     equipamento.setDataUltimaManutencao(txtDataUltimaManutencao.getText());
-     equipamento.setManutencao(cbManutencao.getPromptText());
-
-     if(!txtID.getText().isEmpty())
+     
+     // AQUI SETAMOS O TEXTO QUE APARECE NO FIELD txtDataUltimaManutencao CASO NAO TENHA DATA
+     if(!txtDataUltimaManutencao.getText().equals("<Sem Manutencao>"))
+        equipamento.setDataUltimaManutencao(txtDataUltimaManutencao.getText());
+     
+     // AQUI VERIFICAMOS SE O STATUS ESTa DISABLE PARA COLOCAR 'NAO' COMO PADRaO
+     if(!cbManutencao.isDisabled())
+        equipamento.setManutencao(cbManutencao.getValue());
+     else
+        equipamento.setManutencao("NAO");
+          
+     // O EQUIPAMENTO SÓ PODE SER ASSIGNADO A UM FUNCIONARIO CASO Ja TENHA SIDO CRIADO
+     // CASO CONTRaRIO PRECISA PRIMEIRO CRIAR O EQUIP PRA DEPOIS ASSIGNAR A UM FUNCIONARIO
+     if(!txtID.getText().isEmpty()){
          equipamento.setId(Integer.parseInt(txtID.getText()));
-        
+     
+        if(null != cbFuncionario.getValue()){
+            EquipamentoUtils eqDAO = new EquipamentoUtils();
+            eqDAO.setFuncionarioUsingEquipmentByName(cbFuncionario.getValue(), Integer.parseInt(txtID.getText()));
+        }
+     }
+             
      return equipamento;
  }
+ 
  
  // INSERE OS DADOS DA TABELA/OBJETO PARA OS CAMPOS
  public void setDataTableObject(Equipamento equipamento){
@@ -257,17 +309,47 @@ public class FXMLEquipamentoController {
      txtDataCompra.setText(equipamento.getDataCompra());
      txtDataUltimaManutencao.setText(equipamento.getDataUltimaManutencao());
      txtID.setText(Integer.toString(equipamento.getId()));
+     cbManutencao.setPromptText(equipamento.getManutencao());
+     cbManutencao.setValue(equipamento.getManutencao());
+     
+     if(null != String.valueOf(equipamento.getIdFuncionario())){
+        EquipamentoUtils eqDAO = new EquipamentoUtils();
+        cbFuncionario.setPromptText(eqDAO.getFuncionarioNameById(equipamento.getId()));
+        cbFuncionario.setValue(eqDAO.getFuncionarioNameById(equipamento.getId()));
+     }
+     else{
+        cbFuncionario.setPromptText("Este equipamento nao esta sendo utilizado no Momento");
+        cbFuncionario.setValue(null);
+     }
+     
+           
  }
  
  public void limpaCampos(){
-     txtNome.setText("");
-     txtMarca.setText("");
-     txtPreco.setText("");
-     txtDataCompra.setText("");
-     txtDataUltimaManutencao.setText("");
-     cbManutencao.setPromptText("");
-     cbFuncionario.setPromptText("");
-     txtID.setText("");
+    txtNome.setText("");
+    txtMarca.setText("");
+    txtPreco.setText("");
+    txtDataCompra.setText("");
+    txtDataUltimaManutencao.setText("");
+    txtID.setText("");
+    cbManutencao.setValue(null);
+    cbManutencao.setPromptText("Status");
+    cbFuncionario.setValue(null);
+    cbFuncionario.setPromptText("Que esta usando no Momento");
+    
+ }
+ 
+ public void readComboBox(){
+     ObservableList<String> funcionariosOList = FXCollections.observableArrayList();
+        FuncionarioUtils funcDAO = new FuncionarioUtils();
+        
+        for(Funcionario funcionario : funcDAO.listarAtivos()){
+            funcionariosOList.add(funcionario.getNome());
+        }
+        
+     cbFuncionario.setItems(funcionariosOList);
+     
+     cbManutencao.setItems(optionsManutencao);
  }
  
  public void readTable(){
@@ -275,7 +357,7 @@ public class FXMLEquipamentoController {
      ObservableList<Equipamento> equipamentosOList = FXCollections.observableArrayList();
      EquipamentoDAO eqDAO = new EquipamentoDAO();
      
-     // Busca no banco todas as informações
+     // Busca no banco todas as informacoes
      for(Equipamento equipamento : eqDAO.listar()){
          equipamentosOList.add(new Equipamento(
              equipamento.getId(),
@@ -284,7 +366,8 @@ public class FXMLEquipamentoController {
              equipamento.getPreco(),
              equipamento.getDataCompra(),
              equipamento.getDataUltimaManutencao(),
-             equipamento.getManutencao()
+             equipamento.getManutencao(),
+             equipamento.getIdFuncionario()
          ));
      }
      
@@ -300,5 +383,17 @@ public class FXMLEquipamentoController {
      // Insere os dados na tabela
      tableEquipamento.setItems(equipamentosOList);
  }
+ 
+    public boolean isManutencaoChangeToYes(Equipamento e){
+        
+        EquipamentoUtils eDAO = new EquipamentoUtils();
+        String manutencaoCompare = eDAO.getManutencaoStatusByEquipmentId(e.getId());
+        
+        // VERIFICAR SE O CAMPO STATUS ESTa MUDANDO DE 'NAO' PARA 'SIM'
+        if(!manutencaoCompare.equals(e.getManutencao()) && e.getManutencao().equals("SIM"))
+            return true;
+        
+        return false;
+    }
  
 }
